Add-501-error-code-[NOT_IMPLEMENTED]-main
At present we are throwing  404 BAD REQUEST code when user is sending valid input but NEF /CRISP are not process the same For E.g. (External ID , MSISDN , IPV4 and IPV6) are valid UE inputs for Qod-API, but NEF can process only (MSISDN and IPV6 ). 
in such case we should throw 501 code (NOT IMPLEMENTED)
So we are proposing 501 error code with this changes/commmit